Hide simulation window in tuning mode
When the slider moves throgh a wide range of values the simulation
window appears on the middle of the screen creating a weird effect. The
aim of this commit is to hide such window when the tuning mode is
active. In case of simulation error, the simulation window is shown.